Moved public API functions into dataswarm.h
